Fix, or rather work around for issue with session - service ticket mappings being left in the database in some cases, making the table grow indefinitely.
The solution is to add a post_delete handler which should remove session
service ticket objects when the session is deleted from the database, when
the management command clearsessions is run if nothing else.

If the session is not mapped to the database this is not enough however, in
which case you'd have to run the new purge_session_service_tickets management
command periodically.

See https://github.com/fjollberg/django-cas2/issues/1